requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
pydantic==2.8.0
python-dotenv==1.0.1
supabase==2.7.0
google-search-results==2.4.2
groq==0.9.0
openpyxl==3.1.5
python-multipart==0.0.9
